Function amazon_sp_orders::apis::orders_v0_api::get_order_regulated_info

pub async fn get_order_regulated_info(
    configuration: &Configuration,
    order_id: &str
) -> Result<GetOrderRegulatedInfoResponse, Error<GetOrderRegulatedInfoError>>

Returns regulated information for the order that you specify. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 0.5 | 30 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values then those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.