Struct amazon_sp_orders::models::order::Order

pub struct Order {

    pub amazon_order_id: String,
    pub seller_order_id: Option<String>,
    pub purchase_date: String,
    pub last_update_date: String,
    pub order_status: OrderStatus,
    pub fulfillment_channel: Option<FulfillmentChannel>,
    pub sales_channel: Option<String>,
    pub order_channel: Option<String>,
    pub ship_service_level: Option<String>,
    pub order_total: Option<Box<Money>>,
    pub number_of_items_shipped: Option<i32>,
    pub number_of_items_unshipped: Option<i32>,
    pub payment_execution_detail: Option<Vec<PaymentExecutionDetailItem>>,
    pub payment_method: Option<PaymentMethod>,
    pub payment_method_details: Option<Vec<String>>,
    pub marketplace_id: Option<String>,
    pub shipment_service_level_category: Option<String>,
    pub easy_ship_shipment_status: Option<EasyShipShipmentStatus>,
    pub cba_displayable_shipping_label: Option<String>,
    pub order_type: Option<OrderType>,
    pub earliest_ship_date: Option<String>,
    pub latest_ship_date: Option<String>,
    pub earliest_delivery_date: Option<String>,
    pub latest_delivery_date: Option<String>,
    pub is_business_order: Option<bool>,
    pub is_prime: Option<bool>,
    pub is_premium_order: Option<bool>,
    pub is_global_express_enabled: Option<bool>,
    pub replaced_order_id: Option<String>,
    pub is_replacement_order: Option<bool>,
    pub promise_response_due_date: Option<String>,
    pub is_estimated_ship_date_set: Option<bool>,
    pub is_sold_by_ab: Option<bool>,
    pub is_iba: Option<bool>,
    pub default_ship_from_location_address: Option<Box<Address>>,
    pub buyer_invoice_preference: Option<BuyerInvoicePreference>,
    pub buyer_tax_information: Option<Box<BuyerTaxInformation>>,
    pub fulfillment_instruction: Option<Box<FulfillmentInstruction>>,
    pub is_ispu: Option<bool>,
    pub is_access_point_order: Option<bool>,
    pub marketplace_tax_info: Option<Box<MarketplaceTaxInfo>>,
    pub seller_display_name: Option<String>,
    pub shipping_address: Option<Box<Address>>,
    pub buyer_info: Option<Box<BuyerInfo>>,
    pub automated_shipping_settings: Option<Box<AutomatedShippingSettings>>,
    pub has_regulated_items: Option<bool>,
    pub electronic_invoice_status: Option<ElectronicInvoiceStatus>,

}

Order : Order information.

Fields
amazon_order_id: String

An Amazon-defined order identifier, in 3-7-7 format.

seller_order_id: Option<String>

A seller-defined order identifier.

purchase_date: String

The date when the order was created.

last_update_date: String

The date when the order was last updated. Note: LastUpdateDate is returned with an incorrect date for orders that were last updated before 2009-04-01.

order_status: OrderStatus

The current order status.

fulfillment_channel: Option<FulfillmentChannel>

Whether the order was fulfilled by Amazon (AFN) or by the seller (MFN).

sales_channel: Option<String>

The sales channel of the first item in the order.

order_channel: Option<String>

The order channel of the first item in the order.

ship_service_level: Option<String>

The shipment service level of the order.

order_total: Option<Box<Money>>number_of_items_shipped: Option<i32>

The number of items shipped.

number_of_items_unshipped: Option<i32>

The number of items unshipped.

payment_execution_detail: Option<Vec<PaymentExecutionDetailItem>>

A list of payment execution detail items.

payment_method: Option<PaymentMethod>

The payment method for the order. This property is limited to Cash On Delivery (COD) and Convenience Store (CVS) payment methods. Unless you need the specific COD payment information provided by the PaymentExecutionDetailItem object, we recommend using the PaymentMethodDetails property to get payment method information.

payment_method_details: Option<Vec<String>>

A list of payment method detail items.

marketplace_id: Option<String>

The identifier for the marketplace where the order was placed.

shipment_service_level_category: Option<String>

The shipment service level category of the order. Possible values: Expedited, FreeEconomy, NextDay, SameDay, SecondDay, Scheduled, Standard.

easy_ship_shipment_status: Option<EasyShipShipmentStatus>cba_displayable_shipping_label: Option<String>

Custom ship label for Checkout by Amazon (CBA).

order_type: Option<OrderType>

The type of the order.

earliest_ship_date: Option<String>

The start of the time period within which you have committed to ship the order. In ISO 8601 date time format. Returned only for seller-fulfilled orders. Note: EarliestShipDate might not be returned for orders placed before February 1, 2013.

latest_ship_date: Option<String>

The end of the time period within which you have committed to ship the order. In ISO 8601 date time format. Returned only for seller-fulfilled orders. Note: LatestShipDate might not be returned for orders placed before February 1, 2013.

earliest_delivery_date: Option<String>

The start of the time period within which you have committed to fulfill the order. In ISO 8601 date time format. Returned only for seller-fulfilled orders.

latest_delivery_date: Option<String>

The end of the time period within which you have committed to fulfill the order. In ISO 8601 date time format. Returned only for seller-fulfilled orders that do not have a PendingAvailability, Pending, or Canceled status.

is_business_order: Option<bool>

When true, the order is an Amazon Business order. An Amazon Business order is an order where the buyer is a Verified Business Buyer.

is_prime: Option<bool>

When true, the order is a seller-fulfilled Amazon Prime order.

is_premium_order: Option<bool>

When true, the order has a Premium Shipping Service Level Agreement. For more information about Premium Shipping orders, see "Premium Shipping Options" in the Seller Central Help for your marketplace.

is_global_express_enabled: Option<bool>

When true, the order is a GlobalExpress order.

replaced_order_id: Option<String>

The order ID value for the order that is being replaced. Returned only if IsReplacementOrder = true.

is_replacement_order: Option<bool>

When true, this is a replacement order.

promise_response_due_date: Option<String>

Indicates the date by which the seller must respond to the buyer with an estimated ship date. Returned only for Sourcing on Demand orders.

is_estimated_ship_date_set: Option<bool>

When true, the estimated ship date is set for the order. Returned only for Sourcing on Demand orders.

is_sold_by_ab: Option<bool>

When true, the item within this order was bought and re-sold by Amazon Business EU SARL (ABEU). By buying and instantly re-selling your items, ABEU becomes the seller of record, making your inventory available for sale to customers who would not otherwise purchase from a third-party seller.

is_iba: Option<bool>

When true, the item within this order was bought and re-sold by Amazon Business EU SARL (ABEU). By buying and instantly re-selling your items, ABEU becomes the seller of record, making your inventory available for sale to customers who would not otherwise purchase from a third-party seller.

default_ship_from_location_address: Option<Box<Address>>buyer_invoice_preference: Option<BuyerInvoicePreference>

The buyer’s invoicing preference. Available only in the TR marketplace.

buyer_tax_information: Option<Box<BuyerTaxInformation>>fulfillment_instruction: Option<Box<FulfillmentInstruction>>is_ispu: Option<bool>

When true, this order is marked to be picked up from a store rather than delivered.

is_access_point_order: Option<bool>

When true, this order is marked to be delivered to an Access Point. The access location is chosen by the customer. Access Points include Amazon Hub Lockers, Amazon Hub Counters, and pickup points operated by carriers.

marketplace_tax_info: Option<Box<MarketplaceTaxInfo>>seller_display_name: Option<String>

The seller’s friendly name registered in the marketplace.

shipping_address: Option<Box<Address>>buyer_info: Option<Box<BuyerInfo>>automated_shipping_settings: Option<Box<AutomatedShippingSettings>>has_regulated_items: Option<bool>

Whether the order contains regulated items which may require additional approval steps before being fulfilled.

electronic_invoice_status: Option<ElectronicInvoiceStatus>
Implementations

impl Order

pub fn new(
amazon_order_id: String,
purchase_date: String,
last_update_date: String,
order_status: OrderStatus
) -> Order

Order information.