Module amazon_sp_orders::apis::orders_v0_api

Enums

GetOrderAddressError

struct for typed errors of method get_order_address

GetOrderBuyerInfoError

struct for typed errors of method get_order_buyer_info

GetOrderError

struct for typed errors of method get_order

GetOrderItemsBuyerInfoError

struct for typed errors of method get_order_items_buyer_info

GetOrderItemsError

struct for typed errors of method get_order_items

GetOrderRegulatedInfoError

struct for typed errors of method get_order_regulated_info

GetOrdersError

struct for typed errors of method get_orders

UpdateVerificationStatusError

struct for typed errors of method update_verification_status

Functions

get_order

Returns the order indicated by the specified order ID. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 0.0167 | 20 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_order_address

Returns the shipping address for the specified order. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 0.0167 | 20 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_order_buyer_info

Returns buyer information for the specified order. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 0.0167 | 20 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_order_items

Returns detailed order item information for the order indicated by the specified order ID. If NextToken is provided, it’s used to retrieve the next page of order items. Note: When an order is in the Pending state (the order has been placed but payment has not been authorized), the getOrderItems operation does not return information about pricing, taxes, shipping charges, gift status or promotions for the order items in the order. After an order leaves the Pending state (this occurs when payment has been authorized) and enters the Unshipped, Partially Shipped, or Shipped state, the getOrderItems operation returns information about pricing, taxes, shipping charges, gift status and promotions for the order items in the order. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 0.5 | 30 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_order_items_buyer_info

Returns buyer information for the order items in the specified order. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 0.5 | 30 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_order_regulated_info

Returns regulated information for the order indicated by the specified order ID. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 0.5 | 30 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_orders

Returns orders created or updated during the time frame indicated by the specified parameters. You can also apply a range of filtering criteria to narrow the list of orders returned. If NextToken is present, that will be used to retrieve the orders instead of other criteria. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 0.0167 | 20 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

update_verification_status

Updates (approves or rejects) the verification status of an order containing regulated products. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 0.5 | 30 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.